Move e2e and unit tests to correct files
