+++ starknet_py/tests/e2e/account/account_test.py
@@ -1,14 +1,8 @@
-from unittest.mock import patch, AsyncMock
-
 import pytest
 
-from starknet_py.constants import FEE_CONTRACT_ADDRESS
 from starknet_py.contract import Contract
-from starknet_py.net import KeyPair
-from starknet_py.net.account.account import Account
+from starknet_py.net.client_models import Call
 from starknet_py.net.gateway_client import GatewayClient
-from starknet_py.net.models import parse_address, StarknetChainId
-from starknet_py.net.networks import TESTNET, MAINNET
 from starknet_py.tests.e2e.fixtures.constants import MAX_FEE
 from starknet_py.transaction_exceptions import TransactionRejectedError
 
@@ -40,32 +34,6 @@ async def test_balance_when_token_specified(account, erc20_contract):
     balance = await account.get_balance(erc20_contract.address)
 
     assert balance == 200
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("net", (TESTNET, MAINNET))
-async def test_get_balance_default_token_address(net):
-    client = GatewayClient(net=net)
-    acc_client = Account(
-        client=client,
-        address="0x123",
-        key_pair=KeyPair(123, 456),
-        chain=StarknetChainId.TESTNET,
-    )
-
-    with patch(
-        "starknet_py.net.gateway_client.GatewayClient.call_contract",
-        AsyncMock(),
-    ) as mocked_call_contract:
-        mocked_call_contract.return_value = [0, 0]
-
-        await acc_client.get_balance()
-
-        call = mocked_call_contract.call_args
-
-    (call,) = call[0]
-
-    assert call.to_addr == parse_address(FEE_CONTRACT_ADDRESS)
 
 
 @pytest.mark.asyncio
@@ -162,3 +130,57 @@ async def test_get_class_hash_at(map_contract, account):
     )
 
     assert class_hash != 0
+
+
+@pytest.mark.asyncio()
+async def test_get_nonce(gateway_account):
+    nonce = await gateway_account.get_nonce()
+
+    assert nonce >= 0  # TODO maybe better test
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize(
+    "calls", [[Call(10, 20, [30])], [Call(10, 20, [30]), Call(40, 50, [60])]]
+)
+async def test_sign_invoke_transaction(gateway_account, calls):
+    signed_tx = await gateway_account.sign_invoke_transaction(calls, max_fee=MAX_FEE)
+
+    assert isinstance(signed_tx.signature, list)
+    assert len(signed_tx.signature) > 0
+    assert signed_tx.max_fee == MAX_FEE
+
+
+@pytest.mark.asyncio
+async def test_sign_declare_transaction(gateway_account, map_compiled_contract):
+    signed_tx = await gateway_account.sign_declare_transaction(
+        map_compiled_contract, max_fee=MAX_FEE
+    )
+
+    assert isinstance(signed_tx.signature, list)
+    assert len(signed_tx.signature) > 0
+    assert signed_tx.max_fee == MAX_FEE
+
+
+@pytest.mark.asyncio
+async def test_sign_deploy_account_transaction(gateway_account):
+    class_hash = 0x1234
+    salt = 0x123
+    calldata = [1, 2, 3]
+    signed_tx = await gateway_account.sign_deploy_account_transaction(
+        class_hash, salt, calldata, max_fee=MAX_FEE
+    )
+
+    assert isinstance(signed_tx.signature, list)
+    assert len(signed_tx.signature) > 0
+    assert signed_tx.max_fee == MAX_FEE
+    assert signed_tx.class_hash == class_hash
+    assert signed_tx.contract_address_salt == salt
+    assert signed_tx.constructor_calldata == calldata
+
+
+def test_contract_raises_on_both_client_and_account(gateway_client, gateway_account):
+    with pytest.raises(ValueError) as exinfo:
+        Contract(address=1234, abi=[], client=gateway_client, account=gateway_account)
+
+    assert "Account and client are mutually exclusive" in str(exinfo.value)
